requirements.txt
attrs==21.2.0
bison==0.1.3
certifi==2021.5.30
chardet==4.0.0
Charm-Crypto==0.50
click==8.0.1
decompose-mrobison==0.0.2
filelock==3.0.12
flex==6.14.1
huggingface-hub==0.0.12
hypothesis==6.14.2
idna==2.10
joblib==1.0.1
jsonpointer==2.1
m4==0.0.0
numpy==1.21.0
packaging==21.0
pyparsing==2.4.7
PyYAML==5.4.1
regex==2021.7.6
requests==2.25.1
rfc3987==1.3.8
sacremoses==0.0.45
six==1.16.0
sortedcontainers==2.4.0
strict-rfc3339==0.7
tokenizers==0.10.3
torch==1.9.0
tqdm==4.61.2
transformers==4.8.2
typing-extensions==3.10.0.0
urllib3==1.26.6
validate-email==1.3
